feat: add unit test.

+++ libs/cli/src/migration-generator/electrons/integer.electron.ts
@@ -4,6 +4,6 @@ import { ElectronFactoryInterface } from '../electron-factory';
 
 export class IntegerElectron implements ElectronFactoryInterface {
   buildSqlSchema(electron: MetadataElectron): string {
-    return `.integer('${electron.name}')`;
+    return `.bigInteger('${electron.name}')`;
   }
 }
